Return Long from the GraphQL count field too
CountEntityDataFetcher declared Integer and returned staticApi.count()
directly, so it narrowed the same way the datastore APIs did. The class is
@CompileStatic, so once count() returns Long this no longer compiles:

    [Static type checking] - Possible loss of precision
    from java.lang.Long to java.lang.Integer

The schema field was typed with typeManager.getType(Integer), and GraphQL's
Int is 32 bits by specification, so counts above Integer.MAX_VALUE were
truncated here as well. Type the field with Long, which the type manager
maps to ExtendedScalars.GraphQLLong.

This changes the published schema from `personCount: Int` to
`personCount: Long`. Queries selecting the field and the JSON they return
are unchanged, but generated clients and schema snapshots need
regenerating, so it is called out in the upgrade notes.

+++ grails-data-graphql/core/src/main/groovy/org/grails/gorm/graphql/Schema.groovy
@@ -381,7 +381,7 @@ class Schema {
                     DataFetcher countFetcher = dataFetcherManager.getReadingFetcher(entity, COUNT).orElse(new CountEntityDataFetcher(entity))
 
                     final String countFieldName = namingConvention.getCount(entity)
-                    final GraphQLOutputType countOutputType = (GraphQLOutputType) typeManager.getType(Integer)
+                    final GraphQLOutputType countOutputType = (GraphQLOutputType) typeManager.getType(Long)
 
                     GraphQLFieldDefinition.Builder queryCount = newFieldDefinition()
                             .name(countFieldName)
